update delete list Course

+++ src/main/java/com/OLearning/entity/Packages.java
@@ -14,15 +14,13 @@ import java.util.List;
     public class Packages {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer packageId;
+    private Long packageId;
     private String packageName;
     private Long price;
     private Integer coursesCreated;
     private Integer duration;
     private Boolean isActive;
-
     @OneToMany(mappedBy = "packages", fetch = FetchType.LAZY)
     private List<BuyPackages> listOfBuyPackage;
-
 }
